requirements.txt
pytesseract
Pillow
jiwer
transformers
accelerate
sentencepiece
